feat: 이의 신청 투표 남은 시간 컴포넌트로 변경
